requirements_transcribe_audio.txt
numpy
tqdm
deepspeech-gpu